+++ mobile/android/android-components/components/feature/prompts/src/main/java/mozilla/components/feature/prompts/PromptFeature.kt
@@ -121,6 +121,8 @@ internal const val FRAGMENT_TAG = "mozac_feature_prompt_dialog"
  * a selectable prompt list of credit card options.
  * @property onManageCreditCards A callback invoked when a user selects "Manage credit cards" from
  * the select credit card prompt.
+ * @property onSelectCreditCard A callback invoked when a user selects a credit card from the
+ * select credit card prompt.
  * @property onNeedToRequestPermissions A callback invoked when permissions
  * need to be requested before a prompt (e.g. a file picker) can be displayed.
  * Once the request is completed, [onPermissionsResult] needs to be invoked.
@@ -140,6 +142,7 @@ class PromptFeature private constructor(
     private val onManageLogins: () -> Unit = {},
     private val creditCardPickerView: SelectablePromptView<CreditCard>? = null,
     private val onManageCreditCards: () -> Unit = {},
+    private val onSelectCreditCard: () -> Unit = {},
     onNeedToRequestPermissions: OnNeedToRequestPermissions
 ) : LifecycleAwareFeature, PermissionsFeature, Prompter, ActivityResultHandler,
     UserInteractionHandler {
@@ -170,6 +173,7 @@ class PromptFeature private constructor(
         onManageLogins: () -> Unit = {},
         creditCardPickerView: SelectablePromptView<CreditCard>? = null,
         onManageCreditCards: () -> Unit = {},
+        onSelectCreditCard: () -> Unit = {},
         onNeedToRequestPermissions: OnNeedToRequestPermissions
     ) : this(
         container = PromptContainer.Activity(activity),
@@ -185,7 +189,8 @@ class PromptFeature private constructor(
         loginPickerView = loginPickerView,
         onManageLogins = onManageLogins,
         creditCardPickerView = creditCardPickerView,
-        onManageCreditCards = onManageCreditCards
+        onManageCreditCards = onManageCreditCards,
+        onSelectCreditCard = onSelectCreditCard
     )
 
     constructor(
@@ -202,6 +207,7 @@ class PromptFeature private constructor(
         onManageLogins: () -> Unit = {},
         creditCardPickerView: SelectablePromptView<CreditCard>? = null,
         onManageCreditCards: () -> Unit = {},
+        onSelectCreditCard: () -> Unit = {},
         onNeedToRequestPermissions: OnNeedToRequestPermissions
     ) : this(
         container = PromptContainer.Fragment(fragment),
@@ -217,7 +223,8 @@ class PromptFeature private constructor(
         loginPickerView = loginPickerView,
         onManageLogins = onManageLogins,
         creditCardPickerView = creditCardPickerView,
-        onManageCreditCards = onManageCreditCards
+        onManageCreditCards = onManageCreditCards,
+        onSelectCreditCard = onSelectCreditCard
     )
 
     @Deprecated("Pass only activity or fragment instead")
@@ -231,6 +238,7 @@ class PromptFeature private constructor(
         onManageLogins: () -> Unit = {},
         creditCardPickerView: SelectablePromptView<CreditCard>? = null,
         onManageCreditCards: () -> Unit = {},
+        onSelectCreditCard: () -> Unit = {},
         onNeedToRequestPermissions: OnNeedToRequestPermissions
     ) : this(
         container = activity?.let { PromptContainer.Activity(it) }
@@ -248,7 +256,8 @@ class PromptFeature private constructor(
         loginPickerView = loginPickerView,
         onManageLogins = onManageLogins,
         creditCardPickerView = creditCardPickerView,
-        onManageCreditCards = onManageCreditCards
+        onManageCreditCards = onManageCreditCards,
+        onSelectCreditCard = onSelectCreditCard
     )
 
     private val filePicker = FilePicker(container, store, customTabId, onNeedToRequestPermissions)
@@ -259,7 +268,15 @@ class PromptFeature private constructor(
 
     @VisibleForTesting(otherwise = PRIVATE)
     internal var creditCardPicker =
-        creditCardPickerView?.let { CreditCardPicker(store, it, onManageCreditCards, customTabId) }
+        creditCardPickerView?.let {
+            CreditCardPicker(
+                store = store,
+                creditCardSelectBar = it,
+                manageCreditCardsCallback = onManageCreditCards,
+                selectCreditCardCallback = onSelectCreditCard,
+                sessionId = customTabId
+            )
+        }
 
     override val onNeedToRequestPermissions
         get() = filePicker.onNeedToRequestPermissions
@@ -343,13 +360,39 @@ class PromptFeature private constructor(
 
     /**
      * Notifies the feature of intent results for prompt requests handled by
-     * other apps like file chooser requests.
+     * other apps like credit card and file chooser requests.
      *
      * @param requestCode The code of the app that requested the intent.
      * @param data The result of the request.
+     * @param resultCode The code of the result.
      */
     override fun onActivityResult(requestCode: Int, data: Intent?, resultCode: Int): Boolean {
+        if (requestCode == PIN_REQUEST) {
+            if (resultCode == Activity.RESULT_OK) {
+                creditCardPicker?.onAuthSuccess()
+            } else {
+                creditCardPicker?.onAuthFailure()
+            }
+
+            return true
+        }
+
         return filePicker.onActivityResult(requestCode, resultCode, data)
+    }
+
+    /**
+     * Notifies the feature that the biometric authentication was completed. It will then
+     * either process or dismiss the prompt request.
+     *
+     * @param isAuthenticated True if the user is authenticated successfully from the biometric
+     * authentication prompt or false otherwise.
+     */
+    fun onBiometricResult(isAuthenticated: Boolean) {
+        if (isAuthenticated) {
+            creditCardPicker?.onAuthSuccess()
+        } else {
+            creditCardPicker?.onAuthFailure()
+        }
     }
 
     /**
@@ -763,6 +806,11 @@ class PromptFeature private constructor(
         }
 
         return result
+    }
+
+    companion object {
+        // The PIN request code
+        const val PIN_REQUEST = 303
     }
 }